+++ libsysinspect/src/mdescr/datapatch.rs
